usb: gadget: Enable loopback function for Android
Added f_loopback and f_sourcesink to android.c
composition driver. Used as a no-op driver to
have usb controller on even when none of the
functions are enabled. This is done to facilitate
vbus detection always on in the kernel.

Parts of the driver are commented out to make it
compatible for the android.c composition driver.
The commented sections will be uncommented if
configfs based composition driver is used instead
of android.c driver.

Bug: 21429947
Change-Id: Id15b637f3f5fda723deb8d0b86cc48d7ba768024
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>