android: drivers: workaround debugfs race in binder
For the entries that binder creates in /d/binder/proc/,
make sure the inode private data for those files is not the
binder_proc*, which has a different lifetime than the debugfs
inode entry. Instead, make the private data for the debugfs
file the process pid which can be used to lookup the binder_proc
entry, if it still exists.

Change-Id: Ice878c171db51ef9a4879c2f9299a2deb873d255
Signed-off-by: Riley Andrews <riandrews@android.com>