Update patch set 2
Patch Set 2:

I don't understand why we would want to do this. If the device is just charging it should not enumerate at all. If it enumerates as the gadget loopback device, won't that cause Windows machines to start looking for drivers for this interface?

Patch-set: 2
Reviewer: Mike Lockwood <1004957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0