Update patch set 2
Patch Set 2:

Seems like the "windows adb driver" gets loaded based on the vendor and the product id... ADB claims the interface based on INTERFACE_CLASS SUBCLASS AND PROTOCOL_ID... I am making sure that CLASS, SUBCLASS and PROTOCOL_ID does not match adb requirements. So the driver gets loaded, but the device does not show up if the CLASS/SUBCLASS/PROTOCOL requirments don't match its expectation... Device manager on windows 7 seems to be happy with this. I am yet to test it on WIN 8. I already tested it on MAC and LINUX.

IF an unclaimed interface is going to be a problem, I am planning to make adb to claim the interface even if the INTERFACE CLASS, SUBCLASS, PROTOCOL does not match its expectations.

I also tried to prevent the device from enumeration but that's going make things diverge from whats there in upstream.

Having a no-op approach seemed cleaner as we have almost everything under our control.

Patch-set: 2