Update patch set 2
Patch Set 2:

I tried the following cases to prevent enumeration:

1. Deny SET_CONFIGURATION Request when in charger mode
        Windows pops up a message stating "problem while installing driver software"

2. Increase the power requirement to max (510ma)
        Windows still pops up an error msg stating that it cannot enumerate as it does not have enough power to supply the device.

3. Deny all the setup requests:
        This breaks accessory mode.

So, Unless we find a device CLASS that emulates a "no-op" device, there is no-way to prevent windows from popping an error msg while enumerating. Its either going to pop an error msg while enumerating or its going to look for a driver that is compatible with the device if we respond to all the setup requests.

Patch-set: 2