Dynamically extend saved command line names and tgids
This patch adds logic to dynamically extend the command line and
tgid arrays. When tracing, especially with sched or binder enabled,
we quickly overrun the array containing the names of processes. This
causes loss of information during tracing. A new debugfs entry has
been added to allow extension before a trace.

Change-id: I4be80133cb7a4f2d980fb27f45f95542211fa997
Signed-off-by: Scott Bauer <bauers@google.com>