Dynamically extend saved command line names and tgids
This patch adds logic to dynmaically extend the command line and
tgid arrays. When tracing, especially with sched or binder enabled,
we quickly overrun the array containing the names of processes. This
causes loss of information during tracing. Once we notice we're
about to lose information we increase the sizes of the arrays.

Signed-off-by: Scott Bauer <bauers@google.com>