ext4 crypto: fix return value for ext4_es_scan()
Between 3.10 and 3.18, the abstraction to scan for objects in the slab
cache which can be freed when the system is under memory pressure
changed.  When I backported the ext4 code from 3.18 to the 3.10
kernel, I didn't get the return value required by the calling
conventions for the scan function correct.

This can cause the memory reclaimer to loop indefinitely when the
system is under memory pressure.

Signed-off-by: "Theodore Ts'o" <tytso@google.com>