cpufreq_stats: fix use of cpufreq_for_each_valid_entry() iterator
commit 7f53705d4ca3 (cpufreq_stats: Adds the fucntionality to load
current values for each frequency for all the cores) introduces a
subtle bug in the use of the cpufreq_for_each_valid_entry() iterator.

For example, there are two occasions of this:

	cpufreq_for_each_valid_entry(pos, table);
		count++;

which has a stray ';' after the iterator, meaning that 'count++' is
always executed.

I found this when debugging a crash on an exynos5250-arndale platform, where
the cpufreq_stats driver was crashing during boot.

Change-Id: I9977f395d5b2e26134fc9d379b12197b73ec62e4
Cc: Ruchi Kandoi <kandoiruchi@google.com>
Signed-off-by: Kevin Hilman <khilman@linaro.org>