net/ipv4: android: don't nuke sockets on TCP_TIME_WAIT
Atempting to nuke sockets in TCP_TIME_WAIT state
(tcp_timewait_sockets) results in spinlock lockups.

The TCP_TIME_WAIT state serves two purposes:

1. prevent delayed segments from one connection being accepted from a
later connection relying on the same connection quadruplet.

2. ensure the remote end has closed the connection

or as succintly described in tcp.c: "timeout to catch resent junk
before entering closed"

The duration of the TIME_WAIT state in Linux is not settable and fixed
to 60 seconds time after which the socket is destroyed. With this
commit, rather than interrupting the time-wait processing state to
nuke the connection we chose to let it die on timeout.

This commit fixes the following use case observed on a TI board
running Android kernel 3.14 and one of the 96 family [1] of boards
(HiKey) running Android kernel 3.18.

1. stablish a wifi connection.
2. stream/play a video with the web client.
3. switch off the router or close the connection while playing.

trace:

[   92.754225] BUG: spinlock lockup suspected on CPU#5, netd/1747
[   92.760095]  lock: 0xffffffc03364cfb0, .magic: 0bd10606, .owner:
[   92.768113] CPU: 5 PID: 1747 Comm: netd Tainted:
[   92.775168] Call trace:
[   92.777625] [<ffffffc000088d24>] dump_backtrace+0x0/0x124
[   92.783027] [<ffffffc000088e58>] show_stack+0x10/0x1c
[   92.788092] [<ffffffc000906208>] dump_stack+0x74/0xb8
[   92.793147] [<ffffffc000101e14>] spin_dump+0x68/0x90
[   92.798122] [<ffffffc000102044>] do_raw_spin_lock+0x178/0x1b0
[   92.803871] [<ffffffc00090ca48>] _raw_spin_lock+0x40/0x50
[   92.809283] [<ffffffc0007f271c>] tcp_nuke_addr+0x2cc/0x3b8
[   92.814779] [<ffffffc00081fea8>] devinet_ioctl+0x44c/0x7a0
[   92.820266] [<ffffffc0008222ec>] inet_ioctl+0x88/0xa8
[   92.825330] [<ffffffc00074d2b4>] sock_do_ioctl+0x30/0x70
[   92.830643] [<ffffffc00074db6c>] sock_ioctl+0x234/0x320
[   92.835880] [<ffffffc000215004>] do_vfs_ioctl+0x350/0x59c
[   92.841279] [<ffffffc0002152d0>] SyS_ioctl+0x80/0x98

[1] https://www.96boards.org/

Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>
[Picked and validated on TI platform]
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>