mm: add defer mechanism to ksm
I observed that it is unlikely for KSM to merge new pages from an area
that has already been scanned twice on Android mobile devices, so it's
a waste of power to continue to scan these areas in high frequency.
In this patch a defer mechanism is introduced which is borrowed from
page compaction to KSM. This defer mechanism can lower the scan
frequency automatically in the above case.

Signed-off-by: Wenwei Tao <wenweitaowenwei@gmail.com>