mm: hide the bitops of identifying hugetlb vm
Hugetlb VMAs are not mergeable, that means a VMA couldn't
have VM_HUGETLB and VM_MERGEABLE been set in the same time.
So we use VM_HUGETLB to indicate new mergeable VMAs.
Because of that a VMA which has VM_HUGETLB been set is a
hugetlb VMA only if it doesn't have VM_MERGEABLE been set
in the same time.

Signed-off-by: Wenwei Tao <wenweitaowenwei@gmail.com>