mutex: Add a delay into the SPIN_ON_OWNER wait loop.
(cherry pick from 84d8ce7e0025cac60a8a379a7ee3e59d640fbc03)
NB: adjusted source and commit to pass checkpatch.pl

On arm systems the spin on owner optimization can intermittently cause a
lockup that's usually as long as the waiting thread's cpu timeslice. The
repeated mutex aquisitions + atomics in a single spinning thread can
completely lock out the owner from releasing the kernel mutex. The
owner needs to acquire a spinlock on the relase path and this spinlock
can share a monitor with the other locks and atomics on the waiter path.
Rate limit the waiter so that the thread releasing the mutex never
is starved.

Signed-off-by: Riley Andrews <riandrews@google.com>
Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 23036902
Change-Id: Ie1b64275a0c6141f94faaf3e63fcbf9b5438140c