Update patch set 2
Patch Set 2:

(3 comments)

What about the MAINTENANCE file update, do we want to declare we are the maintainers of these three new files, or let it fall through to the usb list?

I have concern over the interpretation of the value argument; either provide guidance (set int64val, not intval, but can read either, 0 through some small number is an integer, otherwise must be a pointer, for example) or expand the object to have a type field, or switch to a structure and make sure all fields are initialized. Maybe provide macros (ala ERR_PTR, but for positive numbers) to access the string or number policy? Also, if the string is an ascii number string, do you want to validate and interpret it down to the int field with simple_strtoull function, or leave it as a string?

Patch-set: 2
Label: Verified=0