Update patch set 3
Patch Set 3:

> @Dmitry Torokhov
 > thanks for your opinion. I'm also looking for way to sync up
 > upperlayer call from wpa_supplicant and lowerlayer call from dongle
 > event (i.e bottom half), but it's not easy to defer one of two as
 > we wanted.
 > here I attached for spinlock syncup in every read/write try.
 > Appreciated any suggestion.

Insun,

As I mentioned we should reduce usage of the global pointer in the code. Once that is done we can better understand the lifetime of the object(s) and we'll be able to figure out how to avoid accessing invalid/stale pointers.

I posted 2 CLs where I started removing use of g_bcm_cfg. Please take a look and see if you can continue removing the usage.

Thanks,
Dmitry

Patch-set: 3