Update patch set 4
Patch Set 4:

@Dimitry - Android Marshmallow has code to get valid channel list from driver and choosing one of them to run soft AP on.

There is an overlay setting to override 2.4 Band channels, but there is no override for 5 GHz Band channels. The chipsets know this better - so the thought was to get the list of channels from the drivers and use it to choose a channel.

If a chip has separate frimware for STA and AP, then drivers can just query it; but if same firmware is used in both modes, or if the list is coming from driver itself; such code is needed.

I thought Broadcom has separate firmwares for STA and AP - and so this code would not be necessary. Is that not the case?

Patch-set: 4
Reviewer: Vinit Deshpande <1022138@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0