[RFC] drivers: staging: binder crash issue
We find that the binder aborts when doing some asynchronous
transferring, e.g. when a phone call comes in.

If there are asynchronous requests in binder system, and new
requests coming, the asynchronous requests may insert into the
new requests queue.

In this scene, the asynchronous request will affect the
corresponding order of the new requests.

So we think that the asynchronous requests should be added to proc
struct instead of thread’s todo list.

Please let me know, if some things wrong in my opinion.

Change-Id: I367a07633b1305a6d6832d456b758e8f3f899942

Signed-off-by: Victor Chong <victor.chong@linaro.org>