Update patch set 1
Patch Set 1:

We use it to ensure that when a network goes away (e.g., the user walks out of wifi range), applications get a fast error notification and can retry on instead of sitting there blocked in read() calls that will never successfully complete.

An Android device without this will likely see applications get stuck on network switches.

This code has been around for a long time; I don't know why it didn't make it upstream. I'd guess that at the time, we either didn't try, or it was rejected because "it was not useful". We can certainly try to upstream it now.

If you have other suggestions on how to do this, happy to entertain those too. It seems like we could do this by adding a write mode to the option to sock_diag interface, and basically have userspace pick and choose which sockets to kill, but we haven't looked at that yet.

Patch-set: 1
Label: Verified=0