Update patch set 3
Patch Set 3:

Even if we do not want to use this random large prctl value for PR_SET_TIMERSLACK_PID for whatever reasons, we still need to make it consistent across different kernel/common versions.

For example: PR_SET_TIMERSLACK_PID in android-3.14 is 41 (consistent with that in platform/system/core), while in android-3.18 it is 43.

Patch-set: 3