Update patch set 2
Patch Set 2:

In order to do that, in this case, we'd need to provide architecture-specific android baseline configs (and modify scripts/kconfig/merge_config.sh accordingly).

Alternatively, we could just set the default ARCH_MMAP_RND_BITS to our desired value, say 16 bits, which would then be automatically made the default for anything based on this, but then that would differ from upstream, for which I think the default must be 8.

Patch-set: 2