Update patch set 2
Patch Set 2:

(8 comments)

dm="0  <sectormax> android-verity 0xfeedface PARTUUID=abcdefg"
One thing that blocks me to create a custom target is the value of <sectormax>.. Isn't that suppose to refer to the total number of data sectors of the device ? The problem is I do not know that total data blocks of the system paritition(root partition) in advance until I have read the metadata from the rootdev. So the data blocks cannot be a command line parameter. Do you see a way around this ?

If we are able to find an answer for the above question then, I believe creating a custom target would involve adding a new "android-verity" target  to drivers/md/dm-verity.c as I would be sharing all the function pointers  except for the ctr(constructor)  with the target_type "verity" ??

Happy to help in ending up on a middle-ground between chrome-os and Android :) lets hope we find one !!

Patch-set: 2