net: diag: Support SOCK_DESTROY for inet sockets.
This passes the SOCK_DESTROY operation to the underlying
protocol diag handler, or returns -EINVAL if that hander does not
define a destroy operation.

Most of this patch is just renaming functions. This is not
strictly necessary, but it would be fairly counterintuitive to
have the code to destroy inet sockets in a function whose name
starts with inet_diag_get.

Change-Id: I32177b7ee1f9b3b6ed84c71ded4b229f42209c8b
Signed-off-by: Lorenzo Colitti <lorenzo@google.com>