Update patch set 1
Patch Set 1:

Please delay committing this series for a week to see if you get good support for upstream in kernel.org. Better to have a cherry-pick here blessed by the community then this, a revert, then a back-port in the history should they request changes IMHO. I will not stand in your way.

Patch-set: 1