Update patch set 3
Patch Set 3:

From reading the code, it looks like the contract is that inet_csk_listen_stop is only ever called a) from process context, after calling lock_sock, or b) if the socket is orphaned and there is guaranteed not to be a process context.

This is backed up by reading the code. It looks like the only callers of inet_csk_listen_stop are:

1. tcp_close, after calling lock_sock. Already covered in this patch.
2. tcp_disconnect, if the socket is in state LISTEN. tcp_disconnect is called by:
  - tcp_close. Already covered in this patch.
  - sk_prot->disconnect. On TCP sockets, this is called by:
    - __inet_stream_connect, which is always called after calling lock_sock
    - inet_shutdown, which calls lock_sock explicitly
    - inet_csk_listen_stop itself (for child sockets)

Since the only callers of inet_csk_listen_stop only call it after calling lock_sock, this patch should be safe and we don't need to call lock_sock_bh.

Patch-set: 3
Label: Verified=0