Update patch set 2
Patch Set 2:

Since it is in linux-next, it is FROMLIST with an url from the archive pointing to the posting.

There is a thin line between UPSTREAM and BACKPORT. If code actually needed to be changed/added or not. define change? If the conflicts were trivial (as in not using any kernel fu skills)? If KISS like spaces or something, UPSTREAM.

Patch-set: 2