Update patch set 1
Patch Set 1:

Yeah this happens in upstream kernels as well. I was also surprised as this is a pretty obvious bug, and I looked at it for a while to see if I missed something. I do have a pretty easy repro scenario that oopses within 30 mins without this patch, and has been running for 6 hours straight now without problems with it.

As to why upstream hasn't seen it, perhaps it's just because the threshold functionality isn't used extensively, or perhaps unregistering is not a very typical usecase.

Looks like it's currently Acked by one of the cgroup maintainers, but not sure if that means it will be taken yet.

Patch-set: 1