Update patch set 1
Patch Set 1: -Code-Review

(1 comment)

We have a _lot_ more work to do on this, the simulated throttled behavior for writes is too simplistic and does not account for caching.

The good news is this CL applies to _all_ of our common kernels cleanly (at least so far) and upstream is definitely a possibility. However, I would never expect upstream to take this modification :-( but never hurts to ask :-) at least they can help me refine the changes correctness or wishlist even if it never gets Signed Off.

Patch-set: 1