drivers:char:random: Add get_random_long() and use in mmap_rnd().
upstream commit: 9e08f57d684ac2f40685f55f659564bfd91a971e added the
ability to choose from a range of values to use for entropy count in
generating the random offset to the mmap_base address to x86 arch.  The
maximum value on this range was set to 32 bits for 64-bit x86 systems,
but this value could be increased further, requiring more than the 32
bits of randomness provided by get_random_int(), as is already possible
in arm64. Add a new function: get_random_long() which more naturally
fits with the mmap usage of get_random_int() but operates exactly the
same as get_random_int().

Also, fix the shifting constant in mmap_rnd() to be an unsigned long so
that values greater than 31 bits generate an appropriate mask without
overflow.  This is especially important on x86, as its shift instruction
uses a 5-bit mask for the shift operand, which meant that any value for
mmap_rnd_bits over 31 acts as a no-op and effectively disables mmap_base
randomization.

Bug: 26963541
Signed-off-by: dcashman@google.com
Change-Id: Iba53d1b7890920f41bf3a422beda88754edad1d1