Revert "staging: android: lowmemorykiller: Fix task_struct leak"
This reverts commit b659f1b99b2549ebc77eb74a1e6ea288f35ae1ed.

After submitting this patch upstream, it was noticed the patch
description and the content are no longer aligned.

The content of that patch basically dropped out when 83dbbdbb38
("android, lowmemorykiller: remove task handoff notifier")
landed upstream in 3.4. Since then its been rebased along being
basically content-less.

So I propose reverting this so we can zap the change in the next
rebase.

Change-Id: Ifcfd1876c17bb938f5df58dee69c23e4ed0c8d83
Signed-off-by: John Stultz <john.stultz@linaro.org>