Revert "sysfs_net_ipv4: Add sysfs-based knobs for controlling TCP window size"
This reverts commit 848b003b6d7b37649250ae44e944366b4ace6c79.

And looking around in the AOSP tree, I can't seem to find any actual
usage of the:
	tcp_wmem_min
	tcp_wmem_def
	tcp_wmem_max
	tcp_rmem_min
	tcp_rmem_def
	tcp_rmem_max
knobs.

Only the init.rc sets the permissions on them, but I don't see
actual users.

So I wanted to propose dropping them. If I missed a usage,
please let me know and I'll withdraw this.

Signed-off-by: John Stultz <john.stultz@linaro.org>