Update patch set 1
Patch Set 1:

Just to clarify, was the objection here due to the confusion around if EVIOCSSUSPENDBLOCK was still in use, or was it something more subtle?
It seems for kernels > 3.5 EVIOCSSUSPENDBLOCK should not be used, so reverting this seems like should be ok.

Patch-set: 1
Reviewer: John Stultz <1010110@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0