selftests: breakpoint_test: don't optimize away test's side effects
breakpoint_test relies on side effects of reads, writes, and function
calls that the compiler may try to optimize away.  Use volatile types
and annotations to keep the compiler from doing that:

(1) trapped should be volatile, or the compiler may reason trapped is
always 0 in the child process and optimize out the check.

(2) nr_tests should also be volatile, or the compiler may reason that
the child never reads from it and optimize out the write.

(3) Reads and writes to dummy_var trigger watchpoints as a side effect,
so they should go through volatile pointers to make sure each access
happens once (and only once).

(4) The dummy_func function calls trigger breakpoints as a side effect,
so the compiler can't be allowed to inline them or optimize them out:
either will cause the child process not to jump to the expected address.

Without these fixes, clang and gcc will generate broken test cases at
-O1 and higher.

Signed-off-by: Greg Hackmann <ghackmann@google.com>