Update patch set 4
Patch Set 4:

Will let this percolate for a few more weeks in kernel.org. This matches the v2 variant currently under consideration.

The outstanding objection is to move this to the block layer; unfortunately the block layer is behind a queue that breaks the accuracy and device-model-specific pliability of this code. I am hoping they accept that explanation.

Patch-set: 4