Update patch set 3
Patch Set 3:

This is a developer choice, but recognizing that this is touching an actively maintained portion of the kernel. The choice is not if, but when it gets merged into upstream.

This patch is to be upstreamed to kernel.org, then this is a good time to do so, hoping for quick turnaround and minimal if any changes. Currently that is blocked on @android.com email account to provide a means for the author to control and isolate upstream activities with Google responsibilities. If it takes too long, then Janis may have no choice but to use the @google.com account.

If the functionality is required today, there is pressure to get this in, then the CL is merged in now to all android trees and merged down to affected devices.

If kernel.org alters it later, then it will need a revert and application of the upstream decision to all. All the time it is not upstream, but downstream in Android, the the Android team will be responsible for any conflict resolution or porting decisions.

Patch-set: 3