Update patch set 3
Patch Set 3: Code-Review+1

> (1 comment)

I've managed to find half an hour to look at this:

I think this patch is correct, because:
 - name_to_dev_t supports the "%u:%u%c" format
 - dm_get_device() also use name_to_dev_t()

I looked at the dm_substitute_devices() function.  First, it focuses
only on device paths starting with "/" which obviously does not work
with PARTUUID=<UUID> format. Second, it is buggy and does not
substitute correctly the device path when the device path is the last
parameter.  For the second point, since it really is a bug, even if it
does not affect me, I've uploaded a patch
(cf. https://android-review.googlesource.com/#/c/221680/).  I'm
wondering how is it working on your side.

For the first point:
- I do not understand why the substitution is restricted this to "/"
  starting device paths
- I do not see the point of such a substitution because
  dm_get_device() is using name_to_dev_t(), and with this patch,
  android_verity_ctr is also using name_to_dev_t().

Patch-set: 3
Reviewer: Jeremy Compostella <1051574@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1