android: binder: move binder global state into context struct.
In the Treble model, we'll be using the kernel
binder driver not only for regular platform IPC,
but also for IPC to HAL implementations, which
will live in their own process in the Treble world.

Since we want to cleanly separate platform IPC
from so-called "vendor IPC", this change moves
all of the global binder state into a separate
struct, which in turn allows us to instantiate
two binder contexts: one for the platform, and
one for the vendor code.

Each context is associated with its own device
node in /dev.

Having two contexts means that we can easily
enforce that a process having access to only
one binder context isn't able to IPC to anything
in the other binder context, simply by limiting
their access to the relevant device node.

It also allows us to have a separate context
manager (servicemanager) for the vendor services.

Change-Id: Ibf48763aec657f20f4cccd3c336607f4e26213c5