Update patch set 2
Patch Set 2: Code-Review+2

This was originally in the bullhead/angler branches. It belongs in common since it is necessary for the android implementation of cpusets, but unfortunately it (and a few others in the binder driver) never made it there from bullhead/angler.

There is a related CL that should also be taken: I3efbae24a1f6384be1e603fb56f0d3baef61d924 (probably a harder upstream sell).

Patch-set: 2
Label: Code-Review=+2