Update patch set 2
Patch Set 2:

Tim- do you recall the rationale for using CAP_SYS_ADMIN for cpuset_allow_attach (instead of using the default method with CAP_SYS_NICE)? I think it was that we didn't want to grant CAP_SYS_NICE to the AM service that manages the cpusets... but it also seems like CAP_SYS_NICE is the more appropriate one to use.

Patch-set: 2