cpusets: add cpuset allow_attach callback
cpusets are a different cgroup subsystem and therefore need the same
allow_attach callback like the core cpu subsystem. Otherwise the
system_server is not allowed to move pids from one group into another on
behalf of those processes.

This fixes permission denied errors when ENABLE_CPUSETS is turned on for
the device.

Change-Id: Ib7940ba8e886be9eb5bfd8a34b3f6921e242280e
Signed-off-by: Christian Poetzsch <christian.potzsch@imgtec.com>