Update patch set 1
Patch Set 1: Code-Review+2

This looks good; it is possible upstream will be unhappy with "yet another CONFIG for sysctl", and will reject it. In this case, I think we should let it go and just set the sysctl to 3 at boot or carry a small patch to switch the default from 1 to 3. Hopefully this will get accepted, though. Perhaps it would be worth mentioning in the changelog that it would be used by both Android and Debian by default.

Patch-set: 1
Reviewer: Kees Cook <1082649@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2