Update patch set 2
Patch Set 2:

This has not made any progress since the end of 2015, it was dropped in priority according to N schedule changes.  At that time, upstream indicated that exposing sid values directly was undesirable but not out of the question, but that performance numbers would be required as part of justification.

Thus, need to implement an acceptable string-version in binder, then perf, then implement sid version and compare to go this way.

Patch-set: 2
Reviewer: Daniel Cashman <1026058@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0