Update patch set 5
Patch Set 5:

Re: ghackmann
"So I'm worried by any change that gets rid that documentation, even temporarily.  I can imagine after enough time passing, someone might not be aware that the fifth parameter to write_buf is *sometimes* an unchecked userspace pointer (depending on the contents of a different parameter) and ends up using it in a dangerous context."

Since the pointer is coming from a write() it has to first traverse the vfs which does access_ok checks on it. So while if PXN ever comes around a fault could occur if someone memcpys instead of copy_* or *_user but in terms of huge security issues it cant happen afaik.

http://lxr.free-electrons.com/source/fs/read_write.c?v=3.10#L438

Patch-set: 5